requirements.txt
python-dotenv==1.2.1
requests==2.31.0
psycopg2-binary==2.9.11